[CIT-309] move thinedge json into separate crate
