+++ mapper/cumulocity/c8y_translator_lib/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2018"
 chrono = "0.4"
 json = "0.12"
 thiserror = "1.0"
+tedge_json = {path = "../../tedge_json"}
 
 [dev-dependencies]
 pretty_assertions = "0.6"
