+++ mapper/tedge_json/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "tedge_json"
+version = "0.1.0"
+authors = ["root"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+chrono = "0.4"
+json = "0.12"
+thiserror = "1.0"
