+++ mapper/tedge_mapper/src/mapper.rs
@@ -1,4 +1,5 @@
 use c8y_translator_lib::CumulocityJson;
+//use tedge_json::ThinEdgeJsonError;
 
 use tokio::task::JoinHandle;
 use tracing::{debug, error, instrument};
@@ -89,7 +90,7 @@ impl Mapper {
         Ok(())
     }
 
-    fn map(input: &[u8]) -> Result<Vec<u8>, c8y_translator_lib::ThinEdgeJsonError> {
+    fn map(input: &[u8]) -> Result<Vec<u8>, tedge_json::ThinEdgeJsonError> {
         CumulocityJson::from_thin_edge_json(input)
     }
 }
